Remove implicit formats from Defaults
`Defaults` is not the right place to expose the implicit formats. Since
this file is usually read by people, we should keep it as clean as
possible and hide implementation details.

This commit does a small cleanup in update too:

* Uses `dependsOn` to show idiomatic ways of expressing our previous
  implementation.
* Hides the JsonKeyFormat in an `UnsafeLibraryManagementCodec` file.
  I must admit that I don't know why we have an extra format for
  ModuleID.

Regarding dependency positions, the meat of the implementation has been
moved to an independent file, as well as the implicit formats. The
implementation of `dependencyPositionsTask` is now kept to a bare
minimum. Docs have been added too.