Move `baseDirectory` and `target`
I've been experiencing a weird error with regard to the scoping of
baseDirectory.

In my process of analyzing what was going wrong, I've realized that
`baesDirectory` and `target` are defined and scoped twice (one in
CorePlugin), the other one in `IvyPlugin` via `paths`.

It took a while to track where these settings were defined, and since
they are **very very** core to a project, I've added them to
`projectCore`.