Enable parallel execution of scripted in the plugin
The change to enable batched and parallel execution for scripted was
done only for the scripted-sbt project. This pull request enables it for
scripted-plugin, so that all sbt plugins in 1.x. can benefit from it.

By default, it configures a number of parallel instances of 1 (though
batch execution is enabled by default). Users can change the number of
parallel sbt hosts running scripted tests via the
`scriptedParallelInstances` setting.